requirements.txt
flask==2.3.2
gunicorn==20.1.0
requests==2.31.0
python-dotenv==1.0.0
certifi==2023.11.17